+++ package.json
@@ -27,7 +27,7 @@
     "@foundryvtt/foundryvtt-cli": "^1.0.2"
   },
   "type": "module",
-  "main": "gulpfile.cjs",
+  "main": "gulpfile.mjs",
   "directories": {
     "lib": "lib"
   },
